Tagged User in Comment
updated the Comment model to include a field called taggedUsers to record users that are tagged

updated the SocialUser model and added a function to get url. (Tried slugify before but works better without it, so i am leaving the tested coded here just in case.)

added templatetags folder in posts for custom filter so that the urls will be clickable in comment

commented out the add_comment function as it is no longer used.

In order to make this work, you will need to migrate..